Added Summaries functionality on BookLog page
implemented the add summaries button, as well as functionality to view current summaries. the interface for adding a summary remembers what is currently typed until the book entry is submitted. it then appends information from new summary to current summary if one is available. if no summary is available, it will instead show 'no summary available'. 

Added conditionals to form entry so that if leave fields blank for a current entry it doesn't mess with what is currently there. the only exception to this is for the finished checkbox, although that doesn't affect too many things

there is still some styling that i would like to do to make the interface feel a little more consistent, but overall i am pretty happy with it